Commit All Changes and Ready To Review
